backend/requirements.txt
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.0
mysqlclient==2.2.0
